refactor: index 파일 절대 경로 수정
