+++ frontend/src/components/ContactUs/ContactUs.tsx
@@ -1,9 +1,8 @@
 import { useState } from 'react';
 
 import { Button, Input, LoadingBall, Modal, Text, Textarea } from '@/components';
-import { useInput, useInputWithValidate, useToggle } from '@/hooks';
+import { useInput, useInputWithValidate, useToggle, useToast } from '@/hooks';
 import { useAuth } from '@/hooks/authentication';
-import { useToast } from '@/hooks/useToast';
 import { validateEmail } from '@/service/validates';
 import { theme } from '@/style/theme';
