+++ frontend/src/contexts/index.ts
@@ -1,3 +1,3 @@
-export { AuthContext, AuthProvider } from '@/contexts/AuthContext';
-export { HeaderContext, HeaderProvider } from '@/contexts/HeaderContext';
-export { ToastContext, ToastProvider } from '@/contexts/ToastContext';
+export { AuthContext, AuthProvider } from './AuthContext';
+export { HeaderContext, HeaderProvider } from './HeaderContext';
+export { ToastContext, ToastProvider } from './ToastContext';
