+++ frontend/src/routes/index.ts
@@ -1,1 +1,2 @@
 export { END_POINTS } from './endPoints';
+export { default as router } from './router';
